requirements.txt
Flask>=2.0.0
numpy>=1.21.0
requests>=2.25.0
python-chess>=1.999